Add kernel class

+++ src/Viscosity.cpp
@@ -1,4 +1,5 @@
 #include "Viscosity.h"
+#include "Kernels.h"
 #include <GL/glut.h>
 #include <gfx/vec2.h>
 #include <math.h>
@@ -17,7 +18,7 @@ void Viscosity::apply(System* s)
 {
   if (this->active)
   {
-    double mu=10;
+    float mu = 10;
     for (Particle *pi : particles) {
         if (pi->rigid || pi->cloth) continue;
         Vec2f viscosityForce = Vec2f(0, 0);
@@ -25,7 +26,7 @@ void Viscosity::apply(System* s)
         vector<Particle*> targets = s->grid.query(pi);
         for (Particle *pj : targets) {
             viscosityForce += mu * pj->mass * (pj->m_Velocity - pi->m_Velocity) / pj->density
-                             * Viscosity::ddW(pi->m_Position - pj->m_Position);
+                             * Viscos::sdW(pi->m_Position - pj->m_Position);
         }
         pi->vForce = viscosityForce;
         pi->m_Force += viscosityForce;
